refactor(agent/surfaces): extract RunStoreFromSpec into shared module
New module github.com/panyam/mcpkit/agent/surfaces holds the
session-store spec -> RunStore mapping (formerly agentchat's private
buildRunStore). agentchat now calls surfaces.RunStoreFromSpec; the
run-store spec tests moved with it. Behavior-preserving; the
tool-result store mapping stays agentchat-local. part of epic 1193.
