+++ agent/surfaces/chat/go.mod
@@ -11,6 +11,7 @@ require (
 	github.com/muesli/termenv v0.16.0
 	github.com/panyam/mcpkit v0.4.0
 	github.com/panyam/mcpkit/agent/host v0.0.0
+	github.com/panyam/mcpkit/agent/surfaces v0.0.0-00010101000000-000000000000
 	github.com/panyam/mcpkit/ext/otel v0.3.1
 	github.com/redis/go-redis/v9 v9.7.0
 	github.com/spf13/cobra v1.10.2
@@ -136,6 +137,8 @@ require (
 replace github.com/panyam/mcpkit => ../../..
 
 replace github.com/panyam/mcpkit/agent => ../..
+
+replace github.com/panyam/mcpkit/agent/surfaces => ..
 
 replace github.com/panyam/mcpkit/ext/auth => ../../../ext/auth
