+++ agent/surfaces/go.mod
@@ -1,0 +1,53 @@
+module github.com/panyam/mcpkit/agent/surfaces
+
+go 1.26.5
+
+require (
+	github.com/panyam/mcpkit/agent v0.0.0
+	github.com/panyam/mcpkit/agent/store/gorm v0.0.0
+	github.com/panyam/mcpkit/agent/store/redis v0.0.0
+	github.com/redis/go-redis/v9 v9.7.0
+	gorm.io/driver/postgres v1.6.0
+	gorm.io/driver/sqlite v1.6.0
+	gorm.io/gorm v1.31.1
+)
+
+require (
+	github.com/bahlo/generic-list-go v0.2.0 // indirect
+	github.com/buger/jsonparser v1.1.2 // indirect
+	github.com/cespare/xxhash/v2 v2.3.0 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
+	github.com/gorilla/websocket v1.5.0 // indirect
+	github.com/invopop/jsonschema v0.14.0 // indirect
+	github.com/jackc/pgpassfile v1.0.0 // indirect
+	github.com/jackc/pgservicefile v0.0.0-20240606120523-5a60cdf6a761 // indirect
+	github.com/jackc/pgx/v5 v5.6.0 // indirect
+	github.com/jackc/puddle/v2 v2.2.2 // indirect
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/jinzhu/now v1.1.5 // indirect
+	github.com/mattn/go-sqlite3 v1.14.22 // indirect
+	github.com/panyam/gocurrent v0.1.1 // indirect
+	github.com/panyam/goutils v0.1.8 // indirect
+	github.com/panyam/mcpkit v0.4.0 // indirect
+	github.com/panyam/servicekit v0.1.3 // indirect
+	github.com/pb33f/ordered-map/v2 v2.3.1 // indirect
+	github.com/pgvector/pgvector-go v0.4.0 // indirect
+	github.com/yosida95/uritemplate/v3 v3.0.2 // indirect
+	go.yaml.in/yaml/v4 v4.0.0-rc.2 // indirect
+	golang.org/x/crypto v0.31.0 // indirect
+	golang.org/x/sync v0.19.0 // indirect
+	golang.org/x/sys v0.43.0 // indirect
+	golang.org/x/text v0.33.0 // indirect
+	golang.org/x/time v0.15.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20260120221211-b8f7ae30c516 // indirect
+	google.golang.org/grpc v1.80.0 // indirect
+	google.golang.org/protobuf v1.36.11 // indirect
+)
+
+replace github.com/panyam/mcpkit => ../..
+
+replace github.com/panyam/mcpkit/agent => ..
+
+replace github.com/panyam/mcpkit/agent/store/gorm => ../store/gorm
+
+replace github.com/panyam/mcpkit/agent/store/redis => ../store/redis
